move is_authorized to decorators.py; separate out authorization_gateway.py
